gst-plugins-bad: Autospec creation for update from version 1.10.1 to version 1.10.2
# GStreamer 1.10 Release Notes

GStreamer 1.10.0 was originally released on 1st November 2016.
The latest bug-fix release in the 1.10 series is [1.10.2](#1.10.2) and was
released on 29 November 2016.

The GStreamer team is proud to announce a new major feature release in the
stable 1.x API series of your favourite cross-platform multimedia framework!

As always, this release is again packed with new features, bug fixes and other
improvements.

See [https://gstreamer.freedesktop.org/releases/1.10/][latest] for the latest
version of this document.